+++ elements/util-data-test/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/util-data-test
 
+## 17.1.2
+
+### Patch Changes
+
+- [`e75137a9eff`](https://bitbucket.org/atlassian/atlassian-frontend/commits/e75137a9eff) - Adding ability to mark which external users need their sources hydrated
+
 ## 17.1.1
 
 ### Patch Changes